UARTでの文字送信処理の修正(一度に256byte以上送信のときの問題修正) UARTの文字バッファカウンタ変数をuint8_tで定義していたのでDMAバッファ最大値のuin16_tで定義し直した もし，16bit以上のバッファを指定したときも__USAT命令でuint16_tの範囲内で処理するようにした 文字終端検出については_write_r関数に入る前に行ってるっぽいことがわかったのでそれ以後の確認を省略するように変更
Doxyfile内のPROJECT_NAMEがShobonLibL0になってたので修正

+++ DeviceDriver/USARTDriver.h
@@ -1,7 +1,7 @@
 /**
  * @file    USARTDriver.h
  * @author  \@y_sharp
- * @date    2016/05/18
+ * @date    2016/11/08
  * @brief   USARTを用いたマルチバッファ送信のドライバです。
  * syscalls.c 内で標準出力の出力先として利用されています。ユーザが利用する必要はありません。
  * ShobonDeviceDriver.hを介してincludeされるので単体でincludeする必要はありません。
@@ -30,7 +30,7 @@ extern uint8_t getAFR(USART_TypeDef* usart);
 extern void setBaudRate(USART_TypeDef* usart,float baudrate);
 
 extern void USART_Init();
-extern int send_str_DMA(char str[],uint16_t size);
+extern int send_str_DMA(char str[],size_t size);
 extern int8_t getUARTBuffer(int no,uint8_t *dat);
 extern void uart_rx_irq_callback_func(int call_uart);
